+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- Added settings
+- Added Settings
+- Added System Tray Icon
+- Added Hidden mode CLI Argument (-h|hidden)
 
 ## [0.9.0] - 2020-01-02